+++ src/pytest_subtests/__init__.py
@@ -1,0 +1,3 @@
+from .plugin import SubTests
+
+__all__ = ["SubTests"]